s/int/long

+++ keywsync.cc
@@ -165,7 +165,7 @@ int main (int argc, char ** argv) {
   nm_db = setup_db (db_path.c_str());
 
 # ifdef HAVE_NOTMUCH_GET_REV
-  unsigned int revision = notmuch_database_get_revision (nm_db);
+  unsigned long revision = notmuch_database_get_revision (nm_db);
   cout << "* db: current revision: " << revision << endl;
 # endif
